+++ examples/dagster_examples/event_pipeline_demo/repository.py
@@ -5,6 +5,5 @@ from .pipelines import event_ingest_pipeline
 
 def define_repo():
     return RepositoryDefinition(
-        name='event_pipeline_demo_repo',
-        pipeline_dict={'event_ingest_pipeline': event_ingest_pipeline},
+        name='event_pipeline_demo_repo', pipeline_defs=[event_ingest_pipeline]
     )