+++ js_modules/dagit/src/plugins/generic.tsx
@@ -23,67 +23,66 @@ export class SidebarComponent extends React.Component<IPluginSidebarProps> {
   };
 
   render() {
-    const metadata = this.props.solid.definition.metadata;
-    const spark = metadata.find(m => m.key === "kind" && m.value === "spark");
-    if (!spark) {
-      return <span />;
-    }
+    const metadata = this.props.solid.definition.metadata.sort((a, b) =>
+      a.key.localeCompare(b.key)
+    );
 
-    const rest = metadata
-      .filter(m => m !== spark)
-      .sort((a, b) => a.key.localeCompare(b.key));
-
-    if (rest.length === 0) {
+    if (metadata.length === 0) {
       return <span />;
     }
 
     return (
-      <Dialog
-        title={`Metadata: ${this.props.solid.name}`}
-        isOpen={this.state.open}
-        onClose={() =>
-          this.setState({
-            open: false
-          })
-        }
-      >
-        <div
-          className={Classes.DIALOG_BODY}
-          style={{
-            maxHeight: 400,
-            overflow: "scroll"
-          }}
+      <div>
+        <Button icon="duplicate" onClick={this.onClick}>
+          View Metadata
+        </Button>
+        <Dialog
+          title={`Metadata: ${this.props.solid.name}`}
+          isOpen={this.state.open}
+          onClose={() =>
+            this.setState({
+              open: false
+            })
+          }
         >
-          <table
-            className="bp3-html-table bp3-html-table-striped"
-            style={{ width: "100%" }}
+          <div
+            className={Classes.DIALOG_BODY}
+            style={{
+              maxHeight: 400,
+              overflow: "scroll"
+            }}
           >
-            <thead>
-              <tr>
-                <th>Key</th>
-                <th>Value</th>
-              </tr>
-            </thead>
-            <tbody>
-              {rest.map(({ key, value }) => (
-                <tr key={key}>
-                  <td>{startCase(key)}</td>
-                  <td>
-                    <code style={{ whiteSpace: "pre-wrap" }}>{value}</code>
-                  </td>
+            <table
+              className="bp3-html-table bp3-html-table-striped"
+              style={{ width: "100%" }}
+            >
+              <thead>
+                <tr>
+                  <th>Key</th>
+                  <th>Value</th>
                 </tr>
-              ))}
-            </tbody>
-          </table>
-        </div>
-        <div className={Classes.DIALOG_FOOTER}>
-          <div className={Classes.DIALOG_FOOTER_ACTIONS}>
-            <Button onClick={() => this.setState({ open: false })}>
-              Close
-            </Button>
+              </thead>
+              <tbody>
+                {metadata.map(({ key, value }) => (
+                  <tr key={key}>
+                    <td>{startCase(key)}</td>
+                    <td>
+                      <code style={{ whiteSpace: "pre-wrap" }}>{value}</code>
+                    </td>
+                  </tr>
+                ))}
+              </tbody>
+            </table>
           </div>
-        </div>
-      </Dialog>
+          <div className={Classes.DIALOG_FOOTER}>
+            <div className={Classes.DIALOG_FOOTER_ACTIONS}>
+              <Button onClick={() => this.setState({ open: false })}>
+                Close
+              </Button>
+            </div>
+          </div>
+        </Dialog>
+      </div>
     );
   }
 }