+++ python_modules/dagster-dask/dagster_dask/engine.py
@@ -89,11 +89,10 @@ class DaskEngine(IEngine):  # pylint: disable=no-init
 
                     # We ensure correctness in sequencing by letting Dask schedule futures and
                     # awaiting dependencies within each step.
-                    dependencies = [
-                        execution_futures_dict[step_input.prev_output_handle.step_key]
-                        for step_input in step.step_inputs
-                        if step_input.prev_output_handle is not None
-                    ]
+                    dependencies = []
+                    for step_input in step.step_inputs:
+                        for key in step_input.dependency_keys:
+                            dependencies.append(execution_futures_dict[key])
 
                     variables = {
                         'executionParams': {